fix test/run.sh for shUnit2

+++ test/run.sh
@@ -1,34 +1,43 @@
 #! /bin/sh
 
-cd `dirname $0`
+TEST_DIR=`dirname $0`
 TOOL=../semi-automatic-text-editor
+
+setUp()
+{
+  PREV_DIR=$PWD
+  cd $TEST_DIR
+  cp a.txt a.txt.bk
+}
+
+tearDown()
+{
+  if [ -f a.txt.bk ]; then
+    mv a.txt.bk a.txt
+  fi
+  cd $PREV_DIR
+}
 
 testExample1()
 {
-  cp a.txt a.txt.bk
   $TOOL a.txt abc def > /dev/null 2>&1
   diff a.txt 1-expected.txt
   assertEquals 0 $?
-  mv a.txt.bk a.txt
 }
 
 testExample5()
 {
-  cp a.txt a.txt.bk
   $TOOL a.txt abc def 2 > /dev/null 2>&1
   diff a.txt 5-expected.txt
   assertEquals 0 $?
-  mv a.txt.bk a.txt
 }
 
 testExample6()
 {
-  cp a.txt a.txt.bk
   $TOOL a.txt cde def -2 > /dev/null 2>&1
   diff a.txt 6-expected.txt
   assertEquals 0 $?
-  mv a.txt.bk a.txt
 }
 
 # load shunit2
-. ./shunit2
+. $TEST_DIR/shunit2
